requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alpaca-trade-api==3.2.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
deprecation==2.1.0
frozenlist==1.8.0
idna==3.11
msgpack==1.0.3
multidict==6.7.1
numpy==2.4.4
packaging==26.0
pandas==3.0.2
propcache==0.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.1
requests==2.33.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2026.1
urllib3==1.26.20
websocket-client==1.9.0
websockets==10.4
yarl==1.23.0
